Encapsulate generator logic in class
